+++ NEWS.md
@@ -1,9 +1,10 @@
-# midnight (development version) 0.1.0.900-904
+# midnight (development version) 0.1.0.900-905
 
+- Renamed main functions (from `mid_surrogate()` to `mid_reg()`).
 - Added `persp.mid()`.
 - Added the "moon" and "moonlit" color themes
 - Added the package document.
-- Updated wrapper fit function `fit_mid_surrogate()` to improve memory efficiency.
+- Updated wrapper fit function `fit_mid_mid()` to improve memory efficiency.
 - Opened pkgdown website and added favicons.
-- Added `mid_surrogate()` and `make_mid_surrogate()`.
+- Added `mid_surrogate()` and `make_mid_mid()`.
 - Created the package hex logo.